+++ packages/api/endpoints/rules.js
@@ -14,11 +14,10 @@ const {
   getKnexClient,
   isCollisionError,
   RulePgModel,
+  RuleSearch,
   translateApiRuleToPostgresRuleRaw,
   translatePostgresRuleToApiRule,
 } = require('@cumulus/db');
-const { Search, getEsClient } = require('@cumulus/es-client/search');
-const { indexRule, deleteRule } = require('@cumulus/es-client/indexer');
 
 const {
   requireApiVersion,
@@ -47,12 +46,11 @@ const log = new Logger({ sender: '@cumulus/api/rules' });
  * @returns {Promise<Object>} the promise of express response object
  */
 async function list(req, res) {
-  const search = new Search(
-    { queryStringParameters: req.query },
-    'rule',
-    process.env.ES_INDEX
+  const dbSearch = new RuleSearch(
+    { queryStringParameters: req.query }
   );
-  const response = await search.query();
+
+  const response = await dbSearch.query();
   return res.send(response);
 }
 
@@ -93,7 +91,6 @@ async function post(req, res) {
   const {
     rulePgModel = new RulePgModel(),
     knex = await getKnexClient(),
-    esClient = await getEsClient(),
   } = req.testContext || {};
 
   let record;
@@ -116,7 +113,6 @@ async function post(req, res) {
       await createRejectableTransaction(knex, async (trx) => {
         const [pgRecord] = await rulePgModel.create(trx, postgresRule);
         record = await translatePostgresRuleToApiRule(pgRecord, knex);
-        await indexRule(esClient, record, process.env.ES_INDEX);
       });
     } catch (innerError) {
       if (isCollisionError(innerError)) {
@@ -143,7 +139,6 @@ async function post(req, res) {
  * @param {object} params.apiRule           - updated API rule
  * @param {object} params.rulePgModel       - @cumulus/db compatible rule module instance
  * @param {object} params.knex              - Knex object
- * @param {object} params.esClient          - Elasticsearch client
  * @returns {Promise<object>} - promise of an express response object.
  */
 async function patchRule(params) {
@@ -153,7 +148,6 @@ async function patchRule(params) {
     apiRule,
     rulePgModel = new RulePgModel(),
     knex = await getKnexClient(),
-    esClient = await getEsClient(),
   } = params;
 
   log.debug(`rules.patchRule oldApiRule: ${JSON.stringify(oldApiRule)}, apiRule: ${JSON.stringify(apiRule)}`);
@@ -172,7 +166,6 @@ async function patchRule(params) {
     const [pgRule] = await rulePgModel.upsert(trx, apiPgRule);
     log.debug(`rules.patchRule pgRule: ${JSON.stringify(pgRule)}`);
     translatedRule = await translatePostgresRuleToApiRule(pgRule, knex);
-    await indexRule(esClient, translatedRule, process.env.ES_INDEX);
   });
 
   log.info(`rules.patchRule translatedRule: ${JSON.stringify(translatedRule)}`);
@@ -198,7 +191,6 @@ async function patch(req, res) {
   const {
     rulePgModel = new RulePgModel(),
     knex = await getKnexClient(),
-    esClient = await getEsClient(),
   } = req.testContext || {};
 
   const { params: { name }, body } = req;
@@ -216,7 +208,7 @@ async function patch(req, res) {
     apiRule.createdAt = oldApiRule.createdAt;
     apiRule = merge(cloneDeep(oldApiRule), apiRule);
 
-    return await patchRule({ res, oldApiRule, apiRule, knex, esClient, rulePgModel });
+    return await patchRule({ res, oldApiRule, apiRule, knex, rulePgModel });
   } catch (error) {
     log.error('Unexpected error when updating rule:', error);
     if (error instanceof RecordDoesNotExist) {
@@ -242,7 +234,6 @@ async function put(req, res) {
   const {
     rulePgModel = new RulePgModel(),
     knex = await getKnexClient(),
-    esClient = await getEsClient(),
   } = req.testContext || {};
 
   const { params: { name }, body } = req;
@@ -272,7 +263,7 @@ async function put(req, res) {
 
     apiRule.createdAt = oldApiRule.createdAt;
 
-    return await patchRule({ res, oldApiRule, apiRule, knex, esClient, rulePgModel });
+    return await patchRule({ res, oldApiRule, apiRule, knex, rulePgModel });
   } catch (error) {
     log.error('Unexpected error when updating rule:', error);
     if (error instanceof RecordDoesNotExist) {
@@ -293,15 +284,10 @@ async function del(req, res) {
   const {
     rulePgModel = new RulePgModel(),
     knex = await getKnexClient(),
-    esClient = await getEsClient(),
   } = req.testContext || {};
 
   const name = (req.params.name || '').replace(/%20/g, ' ');
-  const esRulesClient = new Search(
-    {},
-    'rule',
-    process.env.ES_INDEX
-  );
+
   let rule;
   let apiRule;
 
@@ -309,26 +295,14 @@ async function del(req, res) {
     rule = await rulePgModel.get(knex, { name });
     apiRule = await translatePostgresRuleToApiRule(rule, knex);
   } catch (error) {
-    // If rule doesn't exist in PG or ES, return not found
     if (error instanceof RecordDoesNotExist) {
-      if (!(await esRulesClient.exists(name))) {
-        log.info('Rule does not exist in Elasticsearch and PostgreSQL');
-        return res.boom.notFound('No record found');
-      }
-      log.info('Rule does not exist in PostgreSQL, it only exists in Elasticsearch. Proceeding with deletion');
-    } else {
-      throw error;
+      return res.boom.notFound('No record found');
     }
+    throw error;
   }
 
   await createRejectableTransaction(knex, async (trx) => {
     await rulePgModel.delete(trx, { name });
-    await deleteRule({
-      esClient,
-      name,
-      index: process.env.ES_INDEX,
-      ignore: [404],
-    });
     if (rule) await deleteRuleResources(knex, apiRule);
   });